Add CMP_EXTRAS
