+++ examples/example_apps/lookandfeel_timeline/CMakeLists.txt
@@ -1,1 +1,3 @@
-add_app_example(lookandfeel_timeline)
+if (${CMP_EXTRAS})
+    add_app_example(lookandfeel_timeline)
+endif()
