+++ source/cmp_lookandfeel.cpp
@@ -13,158 +13,11 @@
 #include <vector>
 
 #include "cmp_datamodels.h"
+#include "cmp_utils.h"
 #include "cmp_graph_line.h"
 #include "cmp_grid.h"
 #include "cmp_label.h"
 #include "juce_core/system/juce_PlatformDefs.h"
-
-/*============================================================================*/
-
-static const std::string getNextCustomLabel(
-    std::vector<std::string>::reverse_iterator& custom_labels_it,
-    const std::vector<std::string>::reverse_iterator& rend) {
-  const auto retval = custom_labels_it++;
-
-  if (retval != rend) {
-    const auto val = *(retval);
-
-    return val;
-  }
-
-  throw std::out_of_range("custom_labels_it is out of range.");
-}
-
-static std::vector<float> getLinearTicks(
-    const std::size_t num_ticks, const cmp::Lim_f lim,
-    const std::vector<float> previous_ticks) {
-  std::vector<float> ticks(num_ticks);
-
-  const auto diff = (lim.max - lim.min) / float(num_ticks);
-  cmp::iota_delta(ticks.begin(), ticks.end(), lim.min + diff / 2.0f, diff);
-
-  return ticks;
-};
-
-static std::vector<float> getLinearTicks_V2(
-    const std::size_t num_ticks_per_power, const cmp::Lim_f lim,
-    const std::vector<float>& previous_ticks) {
-  if (!lim) return {};
-
-  if (!(previous_ticks.empty() || previous_ticks.front() > lim.min ||
-        previous_ticks.back() < lim.max)) {
-    return previous_ticks;
-  }
-
-  const auto delta_min_max = lim.max - lim.min;
-  const auto max_distance = delta_min_max / num_ticks_per_power;
-  const auto log_max_distance = log10(max_distance);
-  const auto log_max_distance_floor = std::floor(log_max_distance);
-  const auto base_value = pow(10.f, log_max_distance_floor);
-  auto multiplier = std::floor(max_distance / base_value);
-  multiplier = multiplier == 0 ? 1 : multiplier;
-  const auto delta = base_value * multiplier;
-  std::vector<float> ticks;
-
-  const auto log_lim_min_floor = std::floor(log10(abs(lim.min)));
-  const auto base_value_min = pow(10.f, log_lim_min_floor);
-  auto multiplier_min = std::floor(lim.min / base_value_min);
-  multiplier_min = multiplier_min == 0 ? 1 : multiplier_min;
-  auto lim_min_round = base_value_min * multiplier_min;
-
-  if (lim_min_round < lim.min - delta) {
-    const auto div = lim.min - delta - lim_min_round;
-    const auto div_round = std::round(div / delta);
-    lim_min_round += div_round * delta;
-  }
-
-  for (float i = lim_min_round - delta * num_ticks_per_power;
-       i <= lim.max + delta * num_ticks_per_power; i += delta) {
-    ticks.push_back(i);
-  }
-
-  return ticks;
-};
-
-static std::pair<float, float> getFirstAndEndFromPreviousTicks(
-    const std::vector<float>& previous_ticks, const cmp::Lim_f lim) {
-  auto start_value = 0.0f;
-  {
-    auto it = previous_ticks.begin();
-    for (; it != previous_ticks.end(); ++it) {
-      if (*it > lim.min) {
-        if (it != previous_ticks.begin() && (it - 1) != previous_ticks.end()) {
-          start_value = *(it - 1);
-        } else {
-          start_value = *it;
-        }
-        break;
-      }
-    }
-  }
-
-  auto end_value = 0.0f;
-  {
-    auto it = previous_ticks.rbegin();
-    for (; it != previous_ticks.rend(); ++it) {
-      if (*it < lim.max) {
-        if (it != previous_ticks.rbegin() &&
-            (it - 1) != previous_ticks.rend()) {
-          end_value = *(it - 1);
-        } else {
-          end_value = *it;
-        }
-        break;
-      }
-    }
-  }
-
-  return {start_value, end_value};
-}
-
-static std::vector<float> getLogarithmicTicks(
-    const std::size_t num_ticks_per_power, const cmp::Lim_f lim,
-    const std::vector<float>& previous_ticks) {
-  if (!lim) return {};
-
-  const auto min_power = log10(lim.min);
-  const auto max_power = log10(lim.max);
-
-  const auto min_power_floor = std::floor(min_power);
-  const auto max_power_ceil = std::ceil(max_power);
-
-  std::vector<float> ticks;
-
-  if (std::abs(max_power - min_power) < 1.0f && !previous_ticks.empty()) {
-    ticks.resize(num_ticks_per_power);
-
-    const auto [start_value, end_value] =
-        getFirstAndEndFromPreviousTicks(previous_ticks, lim);
-
-    auto delta = (end_value - start_value) / num_ticks_per_power;
-    cmp::iota_delta(ticks.begin(), ticks.end(), lim.min, delta);
-
-    return ticks;
-  }
-
-  for (float curr_power = min_power_floor; curr_power < max_power_ceil;
-       ++curr_power) {
-    const auto curr_pos_base = pow(10.f, curr_power);
-
-    const auto delta =
-        pow(10.f, curr_power + 1.f) / static_cast<float>(num_ticks_per_power);
-
-    for (float i = 0; i < num_ticks_per_power; ++i) {
-      const auto tick =
-          floor((curr_pos_base + i * delta) / curr_pos_base) * curr_pos_base;
-
-      ticks.push_back(tick);
-    }
-  }
-
-  return ticks;
-};
-
-/*============================================================================*/
 
 namespace cmp {
 
@@ -1168,248 +1021,5 @@ void PlotLookAndFeel::updateXYTitleLabels(
 }
 
 bool PlotLookAndFeel::isXAxisLabelsBelowGraph() const noexcept { return true; }
-
-////////////////////////////////////////////////////////////////////////////////
-///////////////////////// PlotLookAndFeelTimeline //////////////////////////////
-////////////////////////////////////////////////////////////////////////////////
-
-juce::Rectangle<int> PlotLookAndFeelTimeline::getGraphBounds(
-    const juce::Rectangle<int> bounds,
-    const juce::Component* const plot_comp) const noexcept {
-  const auto estimated_grid_label_width = getGridLabelFont().getStringWidth(
-      std::string(getMaximumAllowedCharacterGridLabel(), 'W'));
-
-  auto graph_bounds = juce::Rectangle<int>();
-
-  if (const auto* plot = dynamic_cast<const Plot*>(plot_comp)) {
-    const auto is_labels_set = getIsLabelsAreSet(plot);
-    const auto [x_grid_label_width, y_grid_label_width] =
-        getMaxGridLabelWidth(plot);
-
-    auto top = getMargin() + getMarginSmall();
-
-    const auto is_x_axis_label_below_graph = isXAxisLabelsBelowGraph();
-    auto bottom =
-        is_x_axis_label_below_graph
-            ? bounds.getHeight() -
-                  (getGridLabelFont().getHeight() + getMargin() +
-                   getXGridLabelDistanceFromGraphBound())
-            : bounds.getHeight() -
-                  (getMargin() + getXGridLabelDistanceFromGraphBound());
-
-    if (is_labels_set.x_label) {
-      bottom -= (getXYTitleFont().getHeight() + getMargin());
-    }
-
-    if (is_labels_set.title_label) {
-      top += getXYTitleFont().getHeight() + getMargin();
-    }
-
-    graph_bounds.setLeft(0);
-    graph_bounds.setTop(int(top));
-    graph_bounds.setRight(bounds.getWidth());
-    graph_bounds.setBottom(bounds.getHeight());
-  }
-
-  return std::move(graph_bounds);
-}
-
-void PlotLookAndFeelTimeline::drawFrame(juce::Graphics& g [[maybe_unused]],
-                                        const juce::Rectangle<int> bounds
-                                        [[maybe_unused]]) {}
-
-void PlotLookAndFeelTimeline::updateGridLabels(
-    const CommonPlotParameterView common_plot_params,
-    const std::vector<GridLine>& grid_lines, StringVector& x_custom_label_ticks,
-    StringVector& y_custom_label_ticks, LabelVector& x_axis_labels_out,
-    LabelVector& y_axis_labels_out) {
-  const auto [x, y, width, height] =
-      getRectangleMeasures<int>(common_plot_params.graph_bounds);
-  const auto font = getGridLabelFont();
-
-  const std::size_t num_horizonal_lines =
-      std::count_if(grid_lines.begin(), grid_lines.end(), [](const auto& gl) {
-        return gl.direction == GridLine::Direction::horizontal;
-      });
-
-  const std::size_t num_vertical_lines =
-      std::count_if(grid_lines.begin(), grid_lines.end(), [](const auto& gl) {
-        return gl.direction == GridLine::Direction::vertical;
-      });
-
-  juce::Rectangle<int>* x_last_label_bound = nullptr;
-  juce::Rectangle<int>* y_last_label_bound = nullptr;
-
-  x_axis_labels_out.clear();
-  y_axis_labels_out.clear();
-
-  const auto use_custom_x_labels = x_custom_label_ticks.size() > 0u;
-  const auto use_custom_y_labels = y_custom_label_ticks.size() > 0u;
-
-  auto custom_x_labels_reverse_it = x_custom_label_ticks.rbegin();
-  if (use_custom_x_labels) {
-    if (x_custom_label_ticks.size() >= num_vertical_lines) {
-      custom_x_labels_reverse_it = x_custom_label_ticks.rbegin() +
-                                   x_custom_label_ticks.size() -
-                                   num_vertical_lines;
-    } else {
-      x_custom_label_ticks.resize(num_vertical_lines);
-
-      custom_x_labels_reverse_it = x_custom_label_ticks.rbegin();
-    }
-  }
-
-  auto custom_y_labels_reverse_it = y_custom_label_ticks.rbegin();
-  if (use_custom_y_labels) {
-    if (y_custom_label_ticks.size() >= num_horizonal_lines) {
-      custom_y_labels_reverse_it = y_custom_label_ticks.rbegin() +
-                                   y_custom_label_ticks.size() -
-                                   num_horizonal_lines;
-    } else {
-      y_custom_label_ticks.resize(num_horizonal_lines);
-
-      custom_y_labels_reverse_it = y_custom_label_ticks.rbegin();
-    }
-  }
-
-  const auto checkInterectionWithLastLabelAndAdd =
-      [&](auto& last_rect, auto& axis_labels, const auto& label,
-          const auto& bound) {
-        if (!last_rect) {
-          axis_labels.push_back({label, bound});
-          last_rect = &axis_labels.back().second;
-        } else {
-          if (!last_rect->intersects(bound)) {
-            axis_labels.push_back({label, bound});
-            last_rect = &axis_labels.back().second;
-          }
-        }
-      };
-
-  const auto getLabelWidthAndHeight =
-      [](const auto& font, const auto& label) -> std::pair<int, int> {
-    return std::make_pair(font.getStringWidth(label),
-                          int(font.getHeightInPoints()));
-  };
-
-  for (auto it = grid_lines.rbegin(); it != grid_lines.rend(); it++) {
-    const auto position = it->position;
-    const auto direction = it->direction;
-    const auto tick = it->tick;
-
-    switch (direction) {
-      case GridLine::Direction::vertical: {
-        const auto label =
-            use_custom_x_labels
-                ? getNextCustomLabel(custom_x_labels_reverse_it,
-                                     x_custom_label_ticks.rend())
-                : valueToString(tick, common_plot_params, true).first;
-
-        const auto [label_width, label_height] =
-            getLabelWidthAndHeight(font, label);
-
-        const auto is_x_axis_label_below_graph = isXAxisLabelsBelowGraph();
-        const auto bound_y =
-            is_x_axis_label_below_graph
-                ? common_plot_params.graph_bounds.getBottom() +
-                      getXGridLabelDistanceFromGraphBound()
-                : common_plot_params.graph_bounds.getTopLeft().y - label_height;
-
-        const auto bound =
-            juce::Rectangle<int>(int(position.x) - label_width / 2, bound_y,
-                                 label_width, label_height);
-
-        checkInterectionWithLastLabelAndAdd(x_last_label_bound,
-                                            x_axis_labels_out, label, bound);
-      } break;
-      case GridLine::Direction::horizontal: {
-        const auto label =
-            use_custom_y_labels
-                ? getNextCustomLabel(custom_y_labels_reverse_it,
-                                     y_custom_label_ticks.rend())
-                : valueToString(tick, common_plot_params, false).first;
-
-        const auto [label_width, label_height] =
-            getLabelWidthAndHeight(font, label);
-
-        const auto bound = juce::Rectangle<int>(
-            x + getMarginSmall(), int(position.y) - label_height / 2,
-            label_width, label_height);
-
-        checkInterectionWithLastLabelAndAdd(y_last_label_bound,
-                                            y_axis_labels_out, label, bound);
-      } break;
-      default:
-        break;
-    }
-  }
-}
-
-bool PlotLookAndFeelTimeline::isXAxisLabelsBelowGraph() const noexcept {
-  return false;
-}
-
-void PlotLookAndFeelTimeline::drawGridLabels(juce::Graphics& g,
-                                             const LabelVector& x_axis_labels,
-                                             const LabelVector& y_axis_labels) {
-  g.setColour(findColour(Plot::x_grid_label_colour));
-  g.setFont(getGridLabelFont());
-  for (const auto& x_axis_text : x_axis_labels) {
-    g.drawText(x_axis_text.first, x_axis_text.second,
-               juce::Justification::centred);
-  }
-  g.setColour(findColour(Plot::y_grid_label_colour));
-  for (const auto& y_axis_text : y_axis_labels) {
-    g.drawText(y_axis_text.first, y_axis_text.second,
-               juce::Justification::centredLeft);
-  }
-}
-
-void PlotLookAndFeelTimeline::updateVerticalGridLineTicksAuto(
-    const juce::Rectangle<int>& bounds,
-    const CommonPlotParameterView& common_plot_parameter_view,
-    const GridType grid_type, const std::vector<float>& previous_ticks,
-    std::vector<float>& x_ticks) noexcept {
-  x_ticks.clear();
-
-  const auto width = bounds.getWidth();
-  const auto height = bounds.getHeight();
-  const auto tiny_grids = grid_type == GridType::small_grid;
-
-  const auto addVerticalTicksLinear = [&]() {
-    std::size_t num_vertical_lines = 5u;
-    if (width > 435u) {
-      num_vertical_lines = 15u;
-    } else if (width <= 435u && width > 175u) {
-      num_vertical_lines = 7u;
-    }
-
-    num_vertical_lines =
-        std::size_t(tiny_grids ? num_vertical_lines * 1.5 : num_vertical_lines);
-
-    x_ticks = getLinearTicks_V2(
-        num_vertical_lines, common_plot_parameter_view.x_lim, previous_ticks);
-  };
-
-  const auto addVerticalTicksLogarithmic = [&]() {
-    std::size_t num_ticks_per_power = 3u;
-    if (width > 435u) {
-      num_ticks_per_power = 10u;
-    } else if (width <= 435u && width > 175u) {
-      num_ticks_per_power = 5u;
-    }
-    num_ticks_per_power = std::size_t(tiny_grids ? num_ticks_per_power * 1.5
-                                                 : num_ticks_per_power);
-
-    x_ticks = getLogarithmicTicks(
-        num_ticks_per_power, common_plot_parameter_view.x_lim, previous_ticks);
-  };
-
-  if (common_plot_parameter_view.x_scaling == Scaling::linear) {
-    addVerticalTicksLinear();
-  } else if (common_plot_parameter_view.x_scaling == Scaling::logarithmic) {
-    addVerticalTicksLogarithmic();
-  }
-}
 
 }  // namespace cmp
